read_local_doc/requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.3.0
bitsandbytes==0.42.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
datasets==4.1.1
dill==0.4.0
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.117.1
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.1
idna==3.10
Jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langsmith==0.4.31
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.2.1
numpy==2.0.2
openai==1.109.1
orjson==3.11.3
packaging==25.0
pandas==2.3.2
peft==0.17.1
propcache==0.3.2
psutil==7.1.0
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
trl==0.23.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.25.0
